requirements.txt
tensorflow
keras
opencv-python_headless
ipykernel
matplotlib
gdown
numpy
pillow
